+++ autodispose-interop/coroutines/Module.md
@@ -1,9 +1,9 @@
-# Module autodispose-courtines-interop
+# Module autodispose-interop-coroutines
 
 Extension functions to interop `ScopeProvider`/`Completable` and `CoroutineScope`, as well as
 `autoDispose(CoroutineScope)` extension functions on RxJava types.
 
-# Package autodispose2.coroutinesinterop
+# Package autodispose2.interop.coroutines
 
 Extension functions to interop `ScopeProvider`/`Completable` and `CoroutineScope`, as well as
 `autoDispose(CoroutineScope)` extension functions on RxJava types.
