+++ autodispose-interop/coroutines/gradle.properties
@@ -15,6 +15,6 @@
 #
 
 POM_NAME=AutoDispose (Coroutines Interop)
-POM_ARTIFACT_ID=autodispose-coroutines-interop
+POM_ARTIFACT_ID=autodispose-interop-coroutines
 POM_PACKAGING=jar
-AUTOMATIC_MODULE_NAME=autodispose2.coroutinesinterop
+AUTOMATIC_MODULE_NAME=autodispose2.interop.coroutines
